feat: move nodebb modules to a separate file, controllers to the lib folder
