+++ lib/controllers.js
@@ -1,14 +1,17 @@
 'use strict';
 
-const groups = require.main.require('./src/groups');
+const { Groups } = require('./nodebb');
 
-const Controllers = module.exports;
+
+const Controllers = {};
 
 Controllers.renderAdminPage = function (req, res, next) {
-	groups.getGroupsFromSet('groups:visible:createtime', 0, -1, (err, groupData) => {
+	Groups.getGroupsFromSet('groups:visible:createtime', 0, -1, (err, groupData) => {
 		if (err) {
 			return next(err);
 		}
 		res.render('admin/plugins/mentions', { groups: groupData });
 	});
 };
+
+module.exports = Controllers;
